+++ geotrek/trekking/serializers.py
@@ -1,3 +1,4 @@
+import copy
 import datetime
 import json
 
@@ -98,7 +99,7 @@ class WebLinkSerializer(TranslatedModelSerializer):
 
 
 class CloseTrekSerializer(TranslatedModelSerializer):
-    category_id = rest_serializers.Field(source='category_id')
+    category_id = rest_serializers.Field(source='prefixed_category_id')
 
     class Meta:
         model = trekking_models.Trek
@@ -131,6 +132,12 @@ class StructureSerializer(rest_serializers.ModelSerializer):
         fields = ('id', 'name')
 
 
+class ChildSerializer(TranslatedModelSerializer):
+    class Meta:
+        model = trekking_models.Trek
+        fields = ('id', )
+
+
 class TrekSerializer(PublishableSerializerMixin, PicturesSerializerMixin,
                      AltimetrySerializerMixin, ZoningSerializerMixin,
                      TranslatedModelSerializer):
@@ -146,6 +153,7 @@ class TrekSerializer(PublishableSerializerMixin, PicturesSerializerMixin,
     relationships = TrekRelationshipSerializer(many=True, source='published_relationships')
     treks = CloseTrekSerializer(many=True, source='published_treks')
     source = RecordSourceSerializer()
+    children = rest_serializers.Field(source='published_children_id')
 
     # Idea: use rest-framework-gis
     parking_location = rest_serializers.SerializerMethodField('get_parking_location')
@@ -162,10 +170,26 @@ class TrekSerializer(PublishableSerializerMixin, PicturesSerializerMixin,
     type2 = TypeSerializer(source='accessibilities', many=True)
     category = rest_serializers.SerializerMethodField('get_category')
 
-    def __init__(self, *args, **kwargs):
-        super(TrekSerializer, self).__init__(*args, **kwargs)
+    def __init__(self, instance=None, *args, **kwargs):
+        # duplicate each trek for each one of its accessibilities
+        if instance and hasattr(instance, '__iter__') and settings.SPLIT_TREKS_CATEGORIES_BY_ACCESSIBILITY:
+            treks = []
+            for trek in instance:
+                treks.append(trek)
+                for accessibility in trek.accessibilities.all():
+                    clone = copy.copy(trek)
+                    clone.accessibility = accessibility
+                    treks.append(clone)
+            instance = treks
+
+        super(TrekSerializer, self).__init__(instance, *args, **kwargs)
 
         from geotrek.tourism import serializers as tourism_serializers
+
+        if settings.SPLIT_TREKS_CATEGORIES_BY_PRACTICE:
+            del self.fields['type1']
+        if settings.SPLIT_TREKS_CATEGORIES_BY_ACCESSIBILITY:
+            del self.fields['type2']
 
         self.fields['information_desks'] = tourism_serializers.InformationDeskSerializer(many=True)
         self.fields['touristic_contents'] = tourism_serializers.CloseTouristicContentSerializer(many=True, source='published_touristic_contents')
@@ -183,7 +207,8 @@ class TrekSerializer(PublishableSerializerMixin, PicturesSerializerMixin,
                   'web_links', 'is_park_centered', 'disabled_infrastructure',
                   'parking_location', 'relationships', 'points_reference',
                   'poi_layer', 'information_desk_layer', 'gpx', 'kml', 'source',
-                  'type1', 'type2', 'category', 'structure', 'treks') + \
+                  'type1', 'type2', 'category', 'structure', 'treks',
+                  'parent', 'children') + \
             AltimetrySerializerMixin.Meta.fields + \
             ZoningSerializerMixin.Meta.fields + \
             PublishableSerializerMixin.Meta.fields + \
@@ -213,13 +238,31 @@ class TrekSerializer(PublishableSerializerMixin, PicturesSerializerMixin,
         return reverse('trekking:trek_kml_detail', kwargs={'pk': obj.pk})
 
     def get_category(self, obj):
-        return {
-            'id': obj.category_id,
-            'label': obj._meta.verbose_name,
-            'type1_label': obj._meta.get_field('practice').verbose_name,
-            'type2_label': obj._meta.get_field('accessibilities').verbose_name,
-            'pictogram': '/static/trekking/trek.svg',
-        }
+        accessibility = getattr(obj, 'accessibility', None)
+        if accessibility:
+            data = {
+                'id': accessibility.prefixed_id,
+                'label': accessibility.name,
+                'pictogram': accessibility.get_pictogram_url(),
+            }
+        elif settings.SPLIT_TREKS_CATEGORIES_BY_PRACTICE and obj.practice:
+            data = {
+                'id': obj.prefixed_category_id,
+                'label': obj.practice.name,
+                'pictogram': obj.practice.get_pictogram_url(),
+            }
+        else:
+            data = {
+                'id': obj.category_id_prefix,
+                'label': obj._meta.verbose_name,
+                'pictogram': '/static/trekking/trek.svg',
+            }
+        data['order'] = settings.TREK_CATEGORY_ORDER
+        if not settings.SPLIT_TREKS_CATEGORIES_BY_PRACTICE:
+            data['type1_label'] = obj._meta.get_field('practice').verbose_name
+        if not settings.SPLIT_TREKS_CATEGORIES_BY_ACCESSIBILITY:
+            data['type2_label'] = obj._meta.get_field('accessibilities').verbose_name
+        return data
 
 
 class POITypeSerializer(PictogramSerializerMixin, TranslatedModelSerializer):